dev/askdayong/requirements.txt
django~=3.0.0
pillow
django-debug-toolbar
django-extesions
django-bootstrap4